RIP RazorLight generator + EnumTMCreator added
